[code setup] (fix) rename old files and delete unsed files
