+++ src/main/cv/python/atNetworktablesTest.py
@@ -1,4 +1,4 @@
-from vision_input import VisionInput
+from visionInput import VisionInput
 from apriltag import AprilTag
 import time
 import ntcore
@@ -41,10 +41,10 @@ while True:
         tagDataPub = table.getDoubleArrayTopic("april_tag_data").publish()
         tagDataPub.set(pose_list)
         
-        # outputStreamPub = table.getDoubleArrayTopic("output_stream").publish()
-        # outputStreamPub.set(annotated_frame.flatten().tolist())
+        outputStreamPub = table.getDoubleArrayTopic("output_stream").publish()
+        outputStreamPub.set(annotated_frame.flatten().tolist())
 
-        #cv2.imshow('result', annotated_frame)
+        cv2.imshow('result', annotated_frame)
         key = cv2.waitKey(1) & 0xFF
         if key == ord('q'):
             break
